// shop-ui/src/app/services/models/product-variant-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProductVariantPropertyValueLinkDto } from '../models/product-variant-property-value-link-dto';
export interface ProductVariantResponseDto {
  demand?: number;
  id?: number;
  image?: string;
  price?: number;
  productId?: number;
  properties?: Array<ProductVariantPropertyValueLinkDto>;
  quantity?: number;
}